% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo.R
\name{bayesianMonteCarloCombine}
\alias{bayesianMonteCarloCombine}
\title{Function that runs combines across Monte Carlo Samples}
\usage{
bayesianMonteCarloCombine(
  discovery,
  steps = 5000,
  burnInSteps = 1000,
  seed = 42,
  definedModel = NULL
)
}
\arguments{
\item{discovery}{Bayesian Meta-Analysis object with Monte Carlo sampling run}

\item{steps}{The number of steps to combine for each sampling}

\item{burnInSteps}{Number of steps for burn-in for each sampling - this should be less than steps}

\item{seedNum}{Number for random seed}

\item{definedMoel}{Model for Bayesian combination - default, basic model}
}
\value{
MetaIntegrator object with monteCarlo results in final results
}
\description{
Function that runs combines across Monte Carlo Samples
}
