% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo.R
\name{bayesianMonteCarloSimulation}
\alias{bayesianMonteCarloSimulation}
\title{Function that runs Bayesian Monte Carlo Sampling}
\usage{
bayesianMonteCarloSimulation(
  discovery,
  numSamplings,
  sampleSize,
  steps = 5000,
  burnInSteps = 1000,
  seed = 42,
  model = NULL
)
}
\arguments{
\item{discovery}{Bayesian Meta-Analysis object with basic bayesian meta-analysis run}

\item{numSamplings}{Number of samplings for monte carlo analysis}

\item{sampleSize}{Size of the samplings for monte carlo}

\item{steps}{The number of steps to combine for each sampling}

\item{burnInSteps}{Number of steps for burn-in for each sampling - this should be less than steps}

\item{model}{Model for Bayesian combination - default, basic model}

\item{seedNum}{Number for random seed}
}
\value{
MetaIntegrator object with monteCarlo populated with samplings
}
\description{
Function that runs Bayesian Monte Carlo Sampling
}
