% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{combineDatasets}
\alias{combineDatasets}
\title{Function that takes individual distributions from getDatasetEffectSizes and combines them}
\usage{
combineDatasets(
  discovery,
  steps = 5000,
  burnIn = 1000,
  seedNum = 42,
  definedModel = NULL
)
}
\arguments{
\item{discovery}{MetaIntegrator object with getDatasetEffectSizes already run.}

\item{steps}{The number of MCMC steps to run for each gene}

\item{seedNum}{Number for random seed}

\item{definedModel}{Optional JAGs model to use for the combine step, if not running standard one}

\item{burnInSteps}{Number of steps for burn-in - this should be less than steps}
}
\value{
discovery$bayesianMeta$finalResults - which contains effect size estimate, tau and Pr(ES = 0) and Rhats
}
\description{
Function that takes individual distributions from getDatasetEffectSizes and combines them
}
