% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFilter.R
\name{filterGenes}
\alias{filterGenes}
\title{Function that returns a filter in discovery based upon the conditions given}
\usage{
filterGenes(
  discovery,
  effectSizeThresh = 0,
  numberStudiesThresh = 2,
  probabilityThresh = 0.05
)
}
\arguments{
\item{discovery}{MetaIntegrator object with combineDatasets already run.}

\item{effectSizeThresh}{The |ES| threshold level}

\item{numberStudiesThresh}{Threshold for the number of studies the gene was observed in}

\item{probabilityThresh}{The probability ES = 0 - lower is more stringent}
}
\value{
discovery object with new filter added
}
\description{
Function that returns a filter in discovery based upon the conditions given
}
