% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleDataset.R
\name{filterRhat}
\alias{filterRhat}
\title{Function that filters out all genes with inidividal dataset Rhats over a certain level}
\usage{
filterRhat(discovery, rhatThresh = 1.1)
}
\arguments{
\item{discovery}{MetaIntegrator object after getSingleDatasetEffectSizes}

\item{rhatThresh}{Your rhat threshold}
}
\value{
MetaIntegrator object with $bayesianMeta$filteredSingleDatasetGenes - describing which genes are removed
}
\description{
Function that filters out all genes with inidividal dataset Rhats over a certain level
}
