% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{getCombineTrace}
\alias{getCombineTrace}
\title{Function that plots the JAGs traces}
\usage{
getCombineTrace(model, parameters = c("MU", "tau"), trace = TRUE)
}
\arguments{
\item{model}{MCMC model from getCombiningModel}

\item{parameters}{Parameters we are interested in - likely MU and tau}

\item{trace}{If true, plots the time series trace. If false, plots the posterior histogram.}
}
\value{
Plot specified by function
}
\description{
Function that plots the JAGs traces
}
