% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{getCombiningModel}
\alias{getCombiningModel}
\title{Function that returns the MCMC model that was used to combine across datasets}
\usage{
getCombiningModel(
  discovery,
  gene,
  steps = 5000,
  burnIn = 1000,
  seedNum = 42,
  definedModel = NULL
)
}
\arguments{
\item{discovery}{MetaIntegrator object with getDatasetEffectSizes already run.}

\item{gene}{The gene we are interested in}

\item{steps}{The number of MCMC steps to run for each gene}

\item{seedNum}{Number for random seed}

\item{definedModel}{Optional JAGs model to use for the combine step, if not running standard one}

\item{burnInSteps}{Number of steps for burn-in - this should be less than steps}
}
\value{
MCMC model for the gene combination
}
\description{
Function that returns the MCMC model that was used to combine across datasets
}
