% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{getGeneTable}
\alias{getGeneTable}
\title{Function that converts probes to gene expression matrix}
\usage{
getGeneTable(dataset, geneSubsetList)
}
\arguments{
\item{dataset}{Dataset in MetaIntegrator format}

\item{geneSubsetList}{A vector with genes that we are interested in -
these will be converted from probe expression values to gene expression values}
}
\value{
MetaIntegrator datasetObject with gene_expr matrix added
}
\description{
Function that converts probes to gene expression matrix
}
