% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{getProbabilityAndFormat}
\alias{getProbabilityAndFormat}
\title{Function that formats the combined data}
\usage{
getProbabilityAndFormat(pooledResultsDetailed)
}
\arguments{
\item{pooledResultsDetailedList}{list of results from combineDatafunction}
}
\value{
Nicely formatted final results
}
\description{
Function that formats the combined data
}
\keyword{internal}
