% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleDataset.R
\name{getSingleDatasetES}
\alias{getSingleDatasetES}
\title{Function that calculates effect size estimate using all cases and controls in a dataset}
\usage{
getSingleDatasetES(dataset, steps = 1000, burnInSteps = 200, seedNum = 42)
}
\arguments{
\item{dataset}{MetaIntegrator dataset with gene_expr populated}

\item{steps}{The number of MCMC steps to run for each gene}

\item{burnInSteps}{Number of steps for burn-in - this should be less than steps}

\item{seedNum}{Number for random seed}
}
\value{
Dataset with datasetResults populated
}
\description{
Function that calculates effect size estimate using all cases and controls in a dataset
}
