% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{makeBayesianForestPlot}
\alias{makeBayesianForestPlot}
\title{Function that creates a forest plot dataset and pooled effect size estimates}
\usage{
makeBayesianForestPlot(discovery, geneToPlot, plotType = "bayesian")
}
\arguments{
\item{discovery}{MetaIntegrator collection of datasets with meta-analysis completed}

\item{geneToPlot}{The gene we are interested in plotting the forest plot for}

\item{plotType}{'bayesian' denotes density forest plot, anything else denotes traditional forest plot}
}
\value{
Forest plot of gene across datasets
}
\description{
Function that creates a forest plot dataset and pooled effect size estimates
}
