% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{probeToGene}
\alias{probeToGene}
\title{Wrapper function that will loop over a MetaIntegrator discovery object and convert probes to genes for all datasets.}
\usage{
probeToGene(discovery, geneSubsetList)
}
\arguments{
\item{discovery}{MetaIntegrator collection of datasets}

\item{geneSubsetList}{A vector with genes that we are interested in -
these will be converted from probe expression values to gene expression values}
}
\value{
MetaIntegrator collection of datasets with gene_expr added to each one.
}
\description{
Wrapper function that will loop over a MetaIntegrator discovery object and convert probes to genes for all datasets.
}
\examples{
probesToGene(discovery, c('BRCA1', 'APOE'))
}
